+++ darjeeling-android-testing/src/main/java/no/agens/darjeeling/android/DarjeelingAndroidTest.kt
@@ -1,9 +1,0 @@
-package no.agens.darjeeling.android
-
-import androidx.appcompat.app.AppCompatActivity
-import kotlin.reflect.KClass
-
-
-@Deprecated("Replaced by DarjeelingActivityTest")
-abstract class DarjeelingAndroidTest<T : AppCompatActivity>(
-    activityClass: KClass<T>) : DarjeelingActivityTest<T>(activityClass)